\name{Interactive Tool for Northsea Macrofauna data}
\alias{webMWTL}
\title{
  Interactive visualisation of the MWTL Northsea macrofauna data
}
\description{
  A shiny interface to see the distribution of macrofauna in the Dutch Northsea.
}

\author{
  Karline Soetaert (R-coding), Anna van der Kaaden (artwork)
}

\usage{
 webMWTL(...)
}

\arguments{
  \item{... }{Arguments passed to the shiny application (not used).
  }
}

\examples{
\dontrun{
  webMWTL()
}  
}

\keyword{ manip }
